refactor for support of other job types
